Added "Google Python Tutorial" under Python List
I found a Google Python Tutorial that reminds me of the HOC that uses tests. It lays out a lot of useful information about python and its well organized so everything is easy to find. It seems to be made for people with a little experience as it doesn't go deep into computer thinking.